refactor : 불필요한 확장성을 위해 만든 RetryableController 삭제
