+++ src/main/java/controller/LadderController.java
@@ -1,5 +1,6 @@
 package controller;
 
+import common.exception.model.IOException;
 import domain.bridge.strategy.RandomBridgeGenerator;
 import domain.ladder.Ladder;
 import domain.ladder.LadderHeight;
@@ -13,11 +14,20 @@ import view.command.Command;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Supplier;
 
-public class LadderController extends RetryableController {
+public class LadderController {
+    public static final int READ_LIMIT = 10;
+    public static final String READ_LIMIT_OVER = String.format("입력 횟수 제한(%d)를 초과하였습니다", READ_LIMIT);
+
+    private final InputView inputView;
+    private final OutputView outputView;
+    private int retryCount;
 
     public LadderController(InputView inputView, OutputView outputView) {
-        super(inputView, outputView);
+        this.inputView = inputView;
+        this.outputView = outputView;
+        this.retryCount = 0;
     }
 
     public Ladder createLadder() {
@@ -31,19 +41,19 @@ public class LadderController extends RetryableController {
         return ladder;
     }
 
-    public void matchPlayerToResult(Ladder ladder) {
+    public void matchPlayerToResult(final Ladder ladder) {
         retry(() -> findPlayerResult(ladder));
         outputView.printEndMessage();
     }
 
-    private void findPlayerResult(Ladder ladder) {
+    private void findPlayerResult(final Ladder ladder) {
         String playerName;
         while (!(playerName = inputView.readPlayerNameForGetResult()).equals(Command.FINISH.getText())) {
             outputView.printPlayerLadderResult(getPlayerLadderResult(ladder, playerName));
         }
     }
 
-    private Map<String, String> getPlayerLadderResult(Ladder ladder, String playerName) {
+    private Map<String, String> getPlayerLadderResult(final Ladder ladder, final String playerName) {
         if (playerName.equals(Command.ALL.getText())) {
             return ladder.findAllPlayersLadderResultValue();
         }
@@ -54,7 +64,7 @@ public class LadderController extends RetryableController {
         return retry(() -> createPlayerNames(inputView.readPlayerNames()));
     }
 
-    private PlayerNames createPlayerNames(String[] playerNamesInput) {
+    private PlayerNames createPlayerNames(final String[] playerNamesInput) {
         List<PlayerName> playerNames = Arrays.stream(playerNamesInput)
                 .map(PlayerName::new)
                 .toList();
@@ -66,7 +76,36 @@ public class LadderController extends RetryableController {
         return retry(() -> new LadderHeight(inputView.readLadderHeight()));
     }
 
-    private LadderResults readLadderResults(int playerCount) {
+    private LadderResults readLadderResults(final int playerCount) {
         return retry(() -> new LadderResults(inputView.readLadderResults(), playerCount));
+    }
+
+    protected <R> R retry(final Supplier<R> supplier) {
+        validateRetryCountLimit();
+        try {
+            R value = supplier.get();
+            retryCount = 0;
+            return value;
+        } catch (Exception exception) {
+            outputView.printErrorMessage(exception.getMessage());
+            return retry(supplier);
+        }
+    }
+
+    protected void retry(final Runnable runnable) {
+        validateRetryCountLimit();
+        try {
+            runnable.run();
+            retryCount = 0;
+        } catch (Exception exception) {
+            outputView.printErrorMessage(exception.getMessage());
+            retry(runnable);
+        }
+    }
+
+    private void validateRetryCountLimit() {
+        if (retryCount++ == READ_LIMIT) {
+            throw new IOException(READ_LIMIT_OVER);
+        }
     }
 }
