+++ src/test/java/controller/RetryableControllerTest.java
@@ -15,15 +15,15 @@ class RetryableControllerTest {
     @Test
     @DisplayName("입력 횟수가 제한 횟수를 초과하면 예외가 발생하고 프로그램이 종료된다")
     void exitApplicationByOverReadLimitCount() {
-        RetryableController retryableController = new RetryableController(new InputView(new Scanner(System.in)), new OutputView());
+        LadderController controller = new LadderController(new InputView(new Scanner(System.in)), new OutputView());
 
-        Assertions.assertThatThrownBy(() -> repeatMethodOverLimitCount(retryableController))
+        Assertions.assertThatThrownBy(() -> repeatMethodOverLimitCount(controller))
                 .isInstanceOf(IOException.class)
-                .hasMessage(RetryableController.READ_LIMIT_OVER);
+                .hasMessage(LadderController.READ_LIMIT_OVER);
     }
 
-    private void repeatMethodOverLimitCount(RetryableController retryableController) {
-        retryableController.retry(() -> {
+    private void repeatMethodOverLimitCount(LadderController controller) {
+        controller.retry(() -> {
             throw new ValidationException("");
         });
     }
